paper-rag project/backend/requirements.txt
flask
flask-cors
pymupdf
sentence-transformers
faiss-cpu
groq
numpy
python-dotenv
